refactor: move the UIRoutes store into the stores/ folder
