+++ src/lib/index.ts
@@ -3,7 +3,7 @@ export { default as GPXLoad } from './GPXLoad.svelte';
 export { default as LocalRoutesList } from './LocalRoutesList.svelte';
 export { default as MapLibreLocalRoutes } from './MapLibreLocalRoutes.svelte';
 export { default as MapLibreWrapper } from './MapLibreWrapper.svelte';
-export { getUIRoutesManager as getUIRoutes } from './db_data/routesData.svelte';
+export { getUIRoutesManager as getUIRoutes } from './stores/routesData.svelte';
 //export { db } from './localDB';
 //export { liveGeoJSONRoutes } from './localDB';
 export type { RouteEntity } from './db_data/localDB';
